Backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
pydantic==2.5.2
pydantic[email]==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
bcrypt==4.1.1
alembic==1.13.0
aiosmtplib==2.0.2
jinja2==3.1.2
requests==2.31.0
